Allow % in powershell source
Elisp's `message` takes a format string as its first argument, meaning it will try to expand anything after a `%`. We want to simply pass through the literal text value, so `(message "%s" full-body)` should be used to avoid the additional expansion.

If this is kept as-is, then any time the user attempts to execute powershell code with a `%` present, emacs presents a user error `Not enough arguments for format string`.